refactor: move required interfaces
