+++ packages/frontend/@n8n/rest-api-client/src/api/tags.ts
@@ -1,0 +1,7 @@
+export interface ITag {
+	id: string;
+	name: string;
+	usageCount?: number;
+	createdAt?: string;
+	updatedAt?: string;
+}
