+++ packages/frontend/editor-ui/src/__tests__/server/fixtures/tags.ts
@@ -1,4 +1,4 @@
-import type { ITag } from '@/Interface';
+import type { ITag } from '@n8n/rest-api-client/api/tags';
 export const tags: ITag[] = [
 	{
 		id: '1',
