+++ packages/frontend/editor-ui/src/__tests__/server/models/tag.ts
@@ -1,4 +1,4 @@
-import type { ITag } from '@/Interface';
+import type { ITag } from '@n8n/rest-api-client/api/tags';
 import { Model } from 'miragejs';
 import type { ModelDefinition } from 'miragejs/-types';
