+++ packages/frontend/editor-ui/src/components/TagsManager/AnnotationTagsManager.ee.vue
@@ -4,7 +4,7 @@ import { useI18n } from '@n8n/i18n';
 import { useToast } from '@/composables/useToast';
 import { useAnnotationTagsStore } from '@/stores/tags.store';
 import TagsManager from './TagsManager.vue';
-import type { ITag } from '@/Interface';
+import type { ITag } from '@n8n/rest-api-client/api/tags';
 import { ANNOTATION_TAGS_MANAGER_MODAL_KEY } from '@/constants';
 
 const i18n = useI18n();
