+++ packages/frontend/editor-ui/src/components/TagsManager/TagsView/TagsView.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { ref, computed } from 'vue';
-import type { ITag, ITagRow } from '@/Interface';
+import type { ITag } from '@n8n/rest-api-client/api/tags';
+import type { ITagRow } from '@/Interface';
 import { useI18n } from '@n8n/i18n';
 import TagsTableHeader from '@/components/TagsManager/TagsView/TagsTableHeader.vue';
 import TagsTable from '@/components/TagsManager/TagsView/TagsTable.vue';
